% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{append_database}
\alias{append_database}
\title{Add new records to existing database}
\usage{
append_database(
  dsn,
  fix_files = NULL,
  device_files = NULL,
  animal_files = NULL,
  deployment_files = NULL
)
}
\arguments{
\item{dsn}{The path to the database file to be created. Currently must be
a .gpkg file.}

\item{fix_files}{A path, or paths, to files with telemetry fixes.}

\item{device_files}{A path, or paths, to CSV files containing information
about the telemetry devices used. This must include an `ID` column.}

\item{animal_files}{A path, or paths, to CSV files containing information
about the animals on which devices were deployed. This must include an `ID`
column.}

\item{deployment_files}{A path, or paths, to CSV files specifying which
devices were deployed on which animals and when. Columns must include:
`AnimalID`, `DeviceID`, `In_Service`, and `Out_Service`.}
}
\value{
[append_database()] returns `TRUE`, invisibly.
}
\description{
Add new records to existing database
}
