% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_animal_fixes}
\alias{get_animal_fixes}
\title{Get fixes tied to animals rather than devices}
\usage{
get_animal_fixes(dsn)
}
\arguments{
\item{dsn}{Path to database}
}
\value{
sf POINT collection
}
\description{
Get fixes tied to animals rather than devices
}
\details{
A wrapper for `sf::st_read()`. Helps simplify code.
}
