% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{get_elevation_dem}
\alias{get_elevation_dem}
\title{Add elevation column for an sf POINT collection.}
\usage{
get_elevation_dem(df, dem_src, units = "meters")
}
\arguments{
\item{df}{an SFC of POINT geometries}

\item{dem_src}{path to Digital Elevation Model file. This must
be a raster file, usually a .tif}

\item{units}{TODO: currently assumes all units are meters.}
}
\description{
Add elevation column for an sf POINT collection.
}
\examples{
\dontrun{
dem_src = "USGS_13_n38w120_20210701.tif"
myDB = system.file("db/telemetry.gpkg", package = "beastr")
}
}
