% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{get_elevation_difference}
\alias{get_elevation_difference}
\title{Get difference between DEM raster and elevation column in an sfc}
\usage{
get_elevation_difference(
  df,
  dem_src,
  elev_field = elevation_gps,
  geom_field = geom,
  abs = TRUE
)
}
\arguments{
\item{df}{spatial datagram with elevation column}

\item{dem_src}{path to Digital Elevation Model file. This must
be a raster file, usually a .tif}

\item{elev_field}{Existing elevation column in `df`}

\item{geom_field}{name of geometry column in df}

\item{abs}{If true, use absolute difference}
}
\description{
Get difference between DEM raster and elevation column in an sfc
}
