% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lotek.R
\name{get_id_from_filename}
\alias{get_id_from_filename}
\title{Extract the id info from a lotek filename}
\usage{
get_id_from_filename(path)
}
\arguments{
\item{path}{path or filename for a lotek data file.}
}
\value{
characters of the device id
}
\description{
Extract the id info from a lotek filename
}
