% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_delims_w_uids}
\alias{read_delims_w_uids}
\title{Read a list of delimited text files that have a unique ID field.
Use this field to pick only distinct entries.}
\usage{
read_delims_w_uids(input_files, id_field = ID)
}
\arguments{
\item{input_files}{a single filename or a list of filenames. These are
delimited text files that share a column name for the unique identifier}

\item{id_field}{column/field name for the unique identifier. Defaults to "ID"}
}
\description{
Read a list of delimited text files that have a unique ID field.
Use this field to pick only distinct entries.
}
