% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lotek.R
\name{read_lotek}
\alias{read_lotek}
\title{Read Lotek gps txt files into a spatial dataframe.}
\usage{
read_lotek(files, remove_duplicates = TRUE, show_col_types = FALSE)
}
\arguments{
\item{files}{Input files. This can be a path to a single file,
a character vector of paths to multiple files, or a list
of paths to multiple files.}

\item{remove_duplicates}{if `FALSE`, duplicate records are included
in the output}

\item{show_col_types}{Passed to readr functions. Useful for debugging.}

\item{id}{the device id #. by default this comes from the filename}

\item{input_crs}{reference system used for interpreting coords from file}

\item{output_crs}{reference system for output spatial df}

\item{tz}{timezone string to convert timezone processing}
}
\description{
This function takes one or more lotek txt files and turns them
into an sf collection of points. If there are multiple versions
of the same fix record, duplicates will be removed by default.
Only the duplicate from the most recent file is kept.
}
\examples{
points = read_lotek(system.file("lotek/33452.txt", package="beastr"))
summary(points)

}
