% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_lotek_2_sf}
\alias{read_lotek_2_sf}
\title{Read a Lotek gps txt file into a spatial dataframe.}
\usage{
read_lotek_2_sf(
  filename,
  id_field = DeviceID,
  id = NULL,
  input_crs = 4326,
  output_crs = 32611,
  tz = "UTC",
  ingest_time = NA
)
}
\arguments{
\item{filename}{a lotek fixed width text file listing fixes from a collar
the filename is assumed to be in the standard lotek format, with device id.}

\item{id}{the device id #. by default this comes from the filename}

\item{input_crs}{reference system used for interpreting coords from file}

\item{output_crs}{reference system for output spatial df}

\item{tz}{timezone string to convert timezone processing}

\item{missed.rm}{if `TRUE`, remove missed fixes}
}
\description{
This function takes a fixed-width delimited lotek txt file and turns it
into an sf collection of points.
}
