% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lotek.R
\name{read_lotek_activity}
\alias{read_lotek_activity}
\title{Read one or more lotek activity text files}
\usage{
read_lotek_activity(files)
}
\arguments{
\item{files}{File paths, either a single file path or a list/vector of
multiple paths. These should point to lotek activity .csv files, which are
not really true CSVs.}
}
\description{
If the same record is contained in multiple files, duplicates are removed.
This happens when a device has multiple downloads with overlapping timespans.
Note that this data is not spatial, but may be linked to corresponding
spatial data through the `device_id` and `time`.
}
\examples{
activity_files = system.file("lotek/activity.csv", package="beastr")
activity = read_lotek_activity(activity_files)
}
