% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{speed}
\alias{speed}
\title{Compute speed from gps telemetry data}
\usage{
speed(
  df,
  id_field = animal_id,
  leadlagfun = dplyr::lead,
  time_field = time,
  geom_field = geom
)
}
\arguments{
\item{df}{Object must also have a time column
returns a vector,  to be used with dplyr::mutate}
}
\description{
Compute speed from gps telemetry data
}
