% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{append_layer}
\alias{append_layer}
\title{Insert rows into a spatial database}
\usage{
append_layer(data, dsn, layer, id_fields = NULL)
}
\arguments{
\item{data}{a tibble or data frame to write to the database. The column names
must be consistent with those in the target layer.}

\item{dsn}{data source name. Typically a path to a geopackage.}

\item{layer}{layer name to append.}

\item{id_fields}{names of field on which to join the new data. These fields
determine whether each row is unique using`dplyr::anti_join()`}
}
\description{
The `append_layer()` method assumes the data source already has a layer
of the same format as the new data. This can work with either spatial or
non-spatial data. It use `sf::st_write()` instead of `DBI::dbAppendTable()`
in order to handle spatial data.
}
