% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastr-package.R
\docType{package}
\name{beastr-package}
\alias{beastr}
\alias{beastr-package}
\title{beastr: Better Environmental/Animal Sensor/Telemetry Repository}
\description{
Manage wildlife telemetry data. This package provides the infrastructure needed to manage wildlife telemetry data. There are functions to read in raw data, build a database that links sensor data to animals, filter bad data, do basic analyses, and export data for use in other applications.
}
\keyword{internal}
