% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_lotek}
\alias{read_lotek}
\title{Read a Lotek gps txt file into a spatial dataframe.}
\usage{
read_lotek(files, id_field = DeviceID, remove_duplicates = TRUE, ...)
}
\arguments{
\item{files}{Input files. This can be a path to a single file,
a character vector of paths to multiple files, or a list
of paths to multiple files.}

\item{id}{the device id #. by default this comes from the filename}

\item{missed.rm}{if `TRUE`, remove missed fixes}

\item{input_crs}{reference system used for interpreting coords from file}

\item{output_crs}{reference system for output spatial df}

\item{tz}{timezone string to convert timezone processing}
}
\description{
This function takes a fixed-width delimited lotek txt file and turns it
into an sf collection of points. If there are records with the same
!!id_field and
}
