% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lotek.R
\name{read_lotek_activity_txt}
\alias{read_lotek_activity_txt}
\title{Read single lotek activity 'csv' with header}
\usage{
read_lotek_activity_txt(filename, show_col_types = FALSE)
}
\arguments{
\item{filename}{A path to a single lotek activity text file.
This should point to a lotek activity .csv files, which are
not really true CSVs.}

\item{show_col_types}{Passed on to [`readr`] functions. Useful for debugging.}
}
\description{
Read single lotek activity 'csv' with header
}
